# src/services/inventory/docs/swagger.yaml
definitions:
...
  dto.ImportOrCreateUserRequest:
    properties:
      ClusterId:
        type: integer
      Create:
        type: boolean
      Password:
        type: string
      Username:
        type: string
    required:
    - ClusterId
    - Username
    type: object
...
  dto.QueueRemoveRequest:
    properties:
      QueueId:
        type: integer
    type: object
  dto.QueueSycronizeRequest:
    properties:
      QueueId:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /{clusterId}/queue/{queueId}:
    get:
      consumes:
      - application/json
      description: Retrieve a queue from cluster
      parameters:
      - description: Cluster id
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Queue id
        in: path
        name: queueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retriave a queue from cluster
      tags:
      - Queue
...
  /{clusterId}/queue/remove:
    delete:
      consumes:
      - application/json
      description: Remove a fila do cluster mas nao altera o cadastro no ostern
      parameters:
      - description: Cluster id
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Request
        in: body
        name: QueueImportRequest
        required: true
        schema:
          $ref: '#/definitions/dto.QueueRemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Remove a fila do cluster
      tags:
      - Queue
  /{clusterId}/queue/syncronize:
    post:
      consumes:
      - application/json
      description: Create a queue in a cluster that not exist in cluster but is registered
        in ostern
      parameters:
      - description: Cluster id
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Request
        in: body
        name: QueueImportRequest
        required: true
        schema:
          $ref: '#/definitions/dto.QueueSycronizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Syncronize a queue between cluster and ostern
      tags:
      - Queue
  /{clusterId}/user:
    get:
      consumes:
      - application/json
      description: Retrieve all users that exist on rabbit cluster. Event if it its
        registered in ostern
      parameters:
      - description: Cluster id from where retrieve users
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retrieve all users from rabbitmq cluster
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new <b>RabbitMQ User mirror</b> from the broker. The user
        must exist in the cluster, the login and hashpassword will be imported
      parameters:
      - description: Request
        in: body
        name: ImportOrCreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ImportOrCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: number
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Syncronize a existing RabbitMQ user from the broker.
      tags:
      - User
  /{clusterId}/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete a mirrored user from the registry, the user will not be
        deleted from the cluster
      parameters:
      - description: User id registered
        in: path
        name: userId
        required: true
        type: integer
      - description: Cluster id from where delete user
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a mirror user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Recovery the details of a specific mirror user that is already
        imported from the cluster
      parameters:
      - description: User id registered
        in: path
        name: userId
        required: true
        type: integer
      - description: Cluster from where the user is
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retrieve a mirror user from broker
      tags:
      - User
  /{clusterId}/user/{userId}/syncronize:
    post:
      consumes:
      - application/json
      description: Cria um ususario que esteja na base do ostern e nao exista no cluster
      parameters:
      - description: Cluster id from where retrieve users
        in: path
        name: clusterId
        required: true
        type: integer
      - description: User id registered
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: number
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Sincronize um ususario no rabbitmq
      tags:
      - User
...
swagger: "2.0"